+++ example/lib/simple_example/simple_example_game.dart
@@ -32,6 +32,10 @@ class SimpleExampleGame extends StatelessWidget {
         },
       ),
       player: MyPlayer(Vector2(140, 140)),
+      cameraConfig: CameraConfig(
+        moveOnlyMapArea: true,
+      ),
+      backgroundColor: Color.fromARGB(255, 10, 53, 89),
     );
   }
 }