+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [2.11.7]
+- Fix diagonal movement speed for enemies
+- Improvements in `moveOnlyMapArea`
+- Adds param `setZoomLimitToFitMap` in `CameraConfig`.
+
 # [2.11.6]
 - Update Flame to `1.5.0`